feat: update restaurant score based on approved dining reviews

+++ src/main/java/com/codeacademy/diningreview/dto/RestaurantResponse.java
@@ -1,16 +1,14 @@
 package com.codeacademy.diningreview.dto;
 
-import java.math.BigDecimal;
-
 public record RestaurantResponse(
         Long id,
         String name,
         String address,
         String zipCode,
         String cuisineType,
-        BigDecimal peanutScore,
-        BigDecimal eggScore,
-        BigDecimal dairyScore,
-        BigDecimal overallScore
+        Double peanutScore,
+        Double eggScore,
+        Double dairyScore,
+        Double overallScore
 ) {
 }
